refactor: extract SettingsJiraExportPane.swift from SettingsIntegrationsPanes.swift
Extracts the SettingsJiraExportPane View (275 lines) — the Settings > Integrations
> Jira Export pane — into its own file.

Byte-identical move; SHA verified against origin/main.

Reduces SettingsIntegrationsPanes.swift from 484 to 205 lines. The
SettingsGitHubExportPane remains in the original file.

Tests: 667 (unchanged).

⚠️ Manual smoke check recommended before merge: render Settings > Integrations
> Jira Export pane and verify layout / behavior unchanged.
